Redux Thunk
In this lesson, you learned how to add Redux middleware such as Redux Logger using the applyMiddleware() function as an argument in Redux's createStore() function. Then you learned how to use Redux Thunk to return a function instead of an action from an action creator, and you updated your app to support loading and failed views when loading campsites data.